add controller_names to motion plan
When constructing the motion plan, the ExecuteTaskSolutionCapability
stores the list of controllers of a SetController stage and sets it in
the succeeding ExecutableTrajectory.

add controller_names to motion plan

When constructing the motion plan, the ExecuteTaskSolutionCapability
stores the list of controllers of a SetController stage and sets it in
the succeeding ExecutableTrajectory.